refactor cache report tests
